move docs file into its own package
