+++ docs/docs_suite_test.go
@@ -1,0 +1,13 @@
+package docs_test
+
+import (
+	. "github.com/onsi/ginkgo"
+	. "github.com/onsi/gomega"
+
+	"testing"
+)
+
+func TestDocs(t *testing.T) {
+	RegisterFailHandler(Fail)
+	RunSpecs(t, "Docs Suite")
+}
